API/godeye/requirements.txt
dotenv==0.0.5
django==4.1.2
djangorestframework==3.14.0
psycopg2==2.9.5
django-cors-headers==3.14.0
drf_spectacular==0.26.1
numpy==1.19.5